refactor!: replace the PNG images for onboarding with WebP for optimal loading time
