+++ apps/web/src/components/onboarding/onboarding-layout.tsx
@@ -1,48 +1,14 @@
-import { Match } from "effect"
 import { AnimatePresence, motion } from "motion/react"
 import type { ReactNode } from "react"
 import { Logo } from "~/components/logo"
 import { Link } from "~/components/ui/link"
+import { getOnboardingImage } from "~/lib/helper/get-onboarding-image"
 
 interface OnboardingLayoutProps {
 	children: ReactNode
 	currentStep?: number
 	totalSteps?: number
 	direction?: "forward" | "backward"
-}
-
-function getOnboardingImage() {
-	const now = new Date()
-	const month = now.getMonth() // 0-11
-	const hour = now.getHours() // 0-23
-
-	// Determine season based on month
-	const season = Match.value(month).pipe(
-		Match.when(
-			(m) => m >= 2 && m <= 4,
-			() => "spring" as const,
-		), // Mar-May
-		Match.when(
-			(m) => m >= 5 && m <= 7,
-			() => "summer" as const,
-		), // Jun-Aug
-		Match.when(
-			(m) => m >= 8 && m <= 10,
-			() => "autumn" as const,
-		), // Sep-Nov
-		Match.orElse(() => "winter" as const), // Dec-Feb
-	)
-
-	// Determine time of day (day = 6AM-6PM, night = 6PM-6AM)
-	const timeOfDay = Match.value(hour).pipe(
-		Match.when(
-			(h) => h >= 6 && h < 18,
-			() => "day" as const,
-		),
-		Match.orElse(() => "night" as const),
-	)
-
-	return `/images/onboarding/${season}-${timeOfDay}.png`
 }
 
 export function OnboardingLayout({
